Add clawpack.geoclaw.met package; deprecate surge as alias shim
Rename the surge Python package to met (the meteorological-forcing name):
move the implementation to src/python/geoclaw/met/ and leave surge/ as thin
re-export shims that emit a single DeprecationWarning and delegate to met.
Add MetData = SurgeData alias in clawpack.geoclaw.data, register both packages
in meson.build, and migrate first-party example/test imports to met. The
surge.data wire filename, rundata.surge_data attr, and Fortran modules are
unchanged; rundata.met_data (clawutil) is deferred. Adds alias/deprecation
tests. Roadmap Decision 1 -> B.
