+++ examples/storm-surge/isaac/setplot.py
@@ -13,8 +13,8 @@ import clawpack.clawutil.data as clawutil
 import clawpack.amrclaw.data as amrclaw
 import clawpack.geoclaw.data as geodata
 import clawpack.geoclaw.util as geoutil
-import clawpack.geoclaw.surge.storm as stormtools
-import clawpack.geoclaw.surge.plot as surgeplot
+import clawpack.geoclaw.met.storm as stormtools
+import clawpack.geoclaw.met.plot as surgeplot
 
 try:
     from setplotfg import setplotfg
