+++ examples/storm-surge/isaac/setplot_kml.py
@@ -25,7 +25,7 @@ import clawpack.clawutil.data as clawutil
 import clawpack.amrclaw.data as amrclaw
 import clawpack.geoclaw.data as geodata
 
-import clawpack.geoclaw.surge.plot as surgeplot
+import clawpack.geoclaw.met.plot as surgeplot
 import clawpack.geoclaw.data as surgedata
 
 try:
